Refactoring: extract Entry classes to entries.py
